Implement backward directory walking async
We use a channel to communicate between the go coroutines, that is, the
main one and the one we spawn to do the directory walking.

Right now, we are only printing the paths that we get, and stop at the
home directory, but this breaks if we invoke `dyncomp` from other
directories that are not under the home directory, say a Volume.